tests: Add a basic test of the daemon
This includes a lot of test infrastructure, such as connecting
to another host, executing commands there, connecting to system
bus there.

Quite complex, but we want to make sure LVM is tested, and never
force ourselves to test it on our own machine.

We upload the udisks-lvm and udisks-lvm-helper commands to the
remote machine, run them there, and connect to the system bus
on the remote machine.

Some complexity due to:
 * GDBusConnection being rather brittle for uses like this
 * The DBus protocol EXTERNAL auth expects UID of user, so
   we have to lookup UID on remote machine in order to auth
   with it.
 * Needed a straight forward way to wait until udisks-lvm is
   initialized, so pattern matching on its stdout.