Manager: Explicitly trigger "change" event after vgcreate.
Otherwise, udev properties might not reflect that the devices are now
physical volumes.  Without this, storaged will not realize that the
LVM configuration has changed and will not discover the new volume
group and the VolumeGroupCreate call will wait forever for it to
appear.

See https://bugzilla.redhat.com/show_bug.cgi?id=1084944.